+++ opensrp-wcaro/src/main/java/org/smartgresiter/wcaro/fragment/FamilyRegisterFragment.java
@@ -4,7 +4,7 @@ import android.support.v7.widget.Toolbar;
 import android.view.View;
 
 import org.smartgresiter.wcaro.R;
-import org.smartgresiter.wcaro.activity.NavigationActivity;
+import org.smartgresiter.wcaro.custom_view.NavigationMenu;
 import org.smartgresiter.wcaro.model.FamilyRegisterFramentModel;
 import org.smartgresiter.wcaro.presenter.FamilyRegisterFragmentPresenter;
 import org.smartregister.family.fragment.BaseFamilyRegisterFragment;
@@ -18,7 +18,7 @@ public class FamilyRegisterFragment extends BaseFamilyRegisterFragment {
         super.setupViews(view);
 
         Toolbar toolbar = view.findViewById(org.smartregister.R.id.register_toolbar);
-        NavigationActivity.getInstance(getActivity(), null, toolbar);
+        NavigationMenu.getInstance(getActivity(), null, toolbar);
 
         View topLeftLayout = view.findViewById(R.id.top_left_layout);
         topLeftLayout.setVisibility(View.GONE);
